fix(Upgrade): Handle dashboard widget metrics migration
